Add diagnostic_lib file
